Improve event handling
This fixes some regressions that occurred, mainly issues with GUI items being able to be stolen through certain click types/actions. It also allows the user to perform certain actions that would normally be blocked if the GUI would allow such changes to it to be made. See the documentation within these changes for specific details.